@@ -13,15 +13,19 @@
 //= require jquery
 //= require jquery_ujs
 //= require turbolinks
-//= require twitter/bootstrap
+//= require make-js-more-like-ruby
+//= require angular
+//= require angular-resource
 //= require jquery-ui
 //= require jquery.ui.all
 //= require jquery.purr
 //= require jquery.ui.datepicker-de
+//= require twitter/bootstrap
 //= require_boxes
 //= require best_in_place
 //= require edit_mode
 //= require slim_breadcrumb
 //= require gmaps4rails
+//= require make-js-more-like-ruby
 //= require your_platform
 //= require_tree .
