@@ -31,8 +31,14 @@ class GroupsController < ApplicationController
     if @group
       @navable = @group
       @title = @group.name
-      @groups = @group.child_groups
-      @users = @group.descendant_users
+
+      @child_groups = @group.child_groups
+      @descendant_users = @group.descendant_users
+      @child_users = @group.child_users
+      @child_users = @child_users.page(params[:page]).per_page(25) # pagination
+
+      user_ids = @group.descendant_users.collect { |user| user.id }
+      @map_address_fields = ProfileField.where( type: "ProfileFieldTypes::Address", profileable_type: "User", profileable_id: user_ids )
     end
   end
 
