@@ -11,7 +11,7 @@
 #
 # It's strongly recommended to check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20130220185631) do
+ActiveRecord::Schema.define(:version => 20130315073719) do
 
   create_table "attachments", :force => true do |t|
     t.string   "file"
@@ -23,6 +23,14 @@ ActiveRecord::Schema.define(:version => 20130220185631) do
     t.datetime "updated_at",   :null => false
     t.string   "content_type"
     t.integer  "file_size"
+  end
+
+  create_table "bookmarks", :force => true do |t|
+    t.integer  "bookmarkable_id"
+    t.string   "bookmarkable_type"
+    t.integer  "user_id"
+    t.datetime "created_at",        :null => false
+    t.datetime "updated_at",        :null => false
   end
 
   create_table "bv_mappings", :force => true do |t|
@@ -60,6 +68,21 @@ ActiveRecord::Schema.define(:version => 20130220185631) do
     t.datetime "created_at",    :null => false
     t.datetime "updated_at",    :null => false
   end
+
+  create_table "geo_locations", :force => true do |t|
+    t.string   "address"
+    t.float    "latitude"
+    t.float    "longitude"
+    t.string   "country"
+    t.string   "country_code"
+    t.string   "city"
+    t.string   "postal_code"
+    t.datetime "created_at",   :null => false
+    t.datetime "updated_at",   :null => false
+    t.datetime "queried_at"
+  end
+
+  add_index "geo_locations", ["address"], :name => "index_geo_locations_on_address"
 
   create_table "groups", :force => true do |t|
     t.string   "name"
@@ -108,14 +131,6 @@ ActiveRecord::Schema.define(:version => 20130220185631) do
     t.integer  "user2_id"
     t.datetime "created_at", :null => false
     t.datetime "updated_at", :null => false
-  end
-
-  create_table "stars", :force => true do |t|
-    t.integer  "starrable_id"
-    t.string   "starrable_type"
-    t.integer  "user_id"
-    t.datetime "created_at",     :null => false
-    t.datetime "updated_at",     :null => false
   end
 
   create_table "status_group_membership_infos", :force => true do |t|
